+++ src/App.tsx
@@ -8,8 +8,8 @@ import ProtectedRoute from "./components/ProtectedRoute";
 import CyberLoadingSpinner from "@/components/CyberLoadingSpinner";
 import NavigationHandler from './components/NavigationHandler';
 const Index = lazy(() => import("./pages/Index"));
-const LandingPage = lazy(() => import("./pages/LandingPage"));
-const LoginPage = lazy(() => import("./pages/LoginPage"));
+const LandingPage = lazy(() => import("./pages/landing/LandingPage"));
+const LoginPage = lazy(() => import("./pages/login/LoginPage"));
 const SignupPage = lazy(() => import("./pages/SignupPage"));
 const ProfileSetupPage = lazy(() => import("./pages/ProfileSetupPage"));
 const MatchingPage = lazy(() => import("./pages/matching/MatchingPage"));
