+++ src/pages/login/LoginPage.tsx
@@ -1,0 +1,13 @@
+import LoginForm from "@/pages/login/components/LoginForm";
+
+const LoginPage = () => (
+  <div className="min-h-screen cyber-grid flex items-center justify-center">
+    <div className="container mx-auto px-4 py-8">
+      <div className="max-w-md mx-auto">
+        <LoginForm />
+        </div>
+    </div>
+  </div>
+);
+
+export default LoginPage;
